Removing unneeded algebraic_jacobian.c
