+++ OMCompiler/SimulationRuntime/c/Makefile.objs
@@ -153,7 +153,6 @@ else
 endif
 ifeq ($(OMC_MINIMAL_RUNTIME),)
   SOLVER_OBJS=$(SOLVER_OBJS_MINIMAL) \
-              analytic_jacobian$(OBJ_EXT) \
               cvode_solver$(OBJ_EXT) \
               dae_mode$(OBJ_EXT) \
               dassl$(OBJ_EXT) \
