--- OMCompiler/SimulationRuntime/c/simulation/solver/analytic_jacobian.c
@@ -1,104 +1,0 @@
-/*
- * This file is part of OpenModelica.
- *
- * Copyright (c) 1998-2022, Open Source Modelica Consortium (OSMC),
- * c/o Linköpings universitet, Department of Computer and Information Science,
- * SE-58183 Linköping, Sweden.
- *
- * All rights reserved.
- *
- * THIS PROGRAM IS PROVIDED UNDER THE TERMS OF THE BSD NEW LICENSE OR THE
- * GPL VERSION 3 LICENSE OR THE OSMC PUBLIC LICENSE (OSMC-PL) VERSION 1.2.
- * ANY USE, REPRODUCTION OR DISTRIBUTION OF THIS PROGRAM CONSTITUTES
- * RECIPIENT'S ACCEPTANCE OF THE OSMC PUBLIC LICENSE OR THE GPL VERSION 3,
- * ACCORDING TO RECIPIENTS CHOICE.
- *
- * The OpenModelica software and the OSMC (Open Source Modelica Consortium)
- * Public License (OSMC-PL) are obtained from OSMC, either from the above
- * address, from the URLs: http://www.openmodelica.org or
- * http://www.ida.liu.se/projects/OpenModelica, and in the OpenModelica
- * distribution. GNU version 3 is obtained from:
- * http://www.gnu.org/copyleft/gpl.html. The New BSD License is obtained from:
- * http://www.opensource.org/licenses/BSD-3-Clause.
- *
- * This program is distributed WITHOUT ANY WARRANTY; without even the implied
- * warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE, EXCEPT AS
- * EXPRESSLY SET FORTH IN THE BY RECIPIENT SELECTED SUBSIDIARY LICENSE
- * CONDITIONS OF OSMC-PL.
- *
- */
-
- /*! \file analytic_jacobian.c
- */
-
-#include "analytic_jacobian.h"
-
-/**
- * @brief Allocate and initialize analytic jacobian.
- *
- * @param sizeCols                  Number of columns of Jacobian
- * @param sizeRows                  Number of rows of Jacobian
- * @param sizeTmpVars               Size of tmp vars array.
- * @param constantEqns              Function pointer for constant equations of Jacobian.
- *                                  NULL if not available.
- * @param sparsePattern             Pointer to sparsity pattern of Jacobian.
- * @return ANALYTIC_JACOBIAN*       Return initialized analytic jacobian.
- */
-ANALYTIC_JACOBIAN* initAnalyticJacobian(unsigned int sizeCols, unsigned int sizeRows, unsigned int sizeTmpVars, int (*constantEqns)(void* data, threadData_t *threadData, void* thisJacobian, void* parentJacobian), SPARSE_PATTERN* sparsePattern) {
-  ANALYTIC_JACOBIAN* jacobian;
-  jacobian = (ANALYTIC_JACOBIAN*) malloc(sizeof(ANALYTIC_JACOBIAN));
-  jacobian->sizeCols = sizeCols;
-  jacobian->sizeRows = sizeRows;
-  jacobian->sizeTmpVars = sizeTmpVars;
-  jacobian->seedVars = (modelica_real*) calloc(sizeCols, sizeof(modelica_real));
-  jacobian->resultVars = (modelica_real*) calloc(sizeRows, sizeof(modelica_real));
-  jacobian->tmpVars = (modelica_real*) calloc(sizeTmpVars, sizeof(modelica_real));
-  jacobian->constantEqns = constantEqns;
-  jacobian->sparsePattern = sparsePattern;
-
-  return jacobian;
-}
-
-/**
- * @brief Copy analytic Jacobian.
- *
- * Sparsity pattern is not copied, only the pointer to it.
- *
- * @param source                  Jacobian that should be copied.
- * @return ANALYTIC_JACOBIAN*     Copy of source.
- */
-ANALYTIC_JACOBIAN* copyAnalyticJacobian(ANALYTIC_JACOBIAN* source) {
-  ANALYTIC_JACOBIAN* jacobian;
-  jacobian = (ANALYTIC_JACOBIAN*) malloc(sizeof(ANALYTIC_JACOBIAN));
-  jacobian->sizeCols = source->sizeCols;
-  jacobian->sizeRows = source->sizeRows;
-  jacobian->sizeTmpVars = source->sizeTmpVars;
-  jacobian->seedVars = (modelica_real*) calloc(source->sizeCols, sizeof(modelica_real));
-  jacobian->resultVars = (modelica_real*) calloc(source->sizeRows, sizeof(modelica_real));
-  jacobian->tmpVars = (modelica_real*) calloc(source->sizeTmpVars, sizeof(modelica_real));
-  jacobian->constantEqns = source->constantEqns;
-  jacobian->sparsePattern = source->sparsePattern;
-
-  return jacobian;
-}
-
-/**
- * @brief Free memory of analytic Jacobian.
- *
- * @param jacobian  Pointer to Jacobian.
- */
-void freeAnalyticJacobian(ANALYTIC_JACOBIAN* jacobian) {
-  if(jacobian==NULL) {
-    return;
-  }
-
-  free(jacobian->seedVars);
-  jacobian->seedVars = NULL;
-  free(jacobian->resultVars);
-  jacobian->resultVars = NULL;
-  free(jacobian->tmpVars);
-  jacobian->tmpVars = NULL;
-
-  free(jacobian);
-  jacobian = NULL;
-}
